add usage example point-at
