+++ public/usage-examples/geometry/point_at/point_at-1-simple.txt
@@ -1,0 +1,3 @@
+#### Create Point
+
+The following code shows examples of using [Point At](/api/geometry/#point-at) to show how to create a point at a specific location and display it as a small circle.
